JNG-4196 support static queries for sdk
